Fixed Lines Sometimes Shifting Around in Output
In situations where some actual code was followed directly by a non-code line, without being separated by an empty line, the non-code line would get printed before the actual code was because of how the logic for dumping the translation buffer was handled. Now, we just ensure that if we run into a non-code line, we attempt to dump the translation buffer, fixing the issue.

Additionally, separated out the logic for dumping the translation buffer into a function since I'm using it in two places now, and also removed the second translationBuffer we used to keep (we can just do parseGeckoCode directly into the outputStream lol, no reason to have that indirection into a secondary buffer).