Added ".RESET" Declaration Detection
Makes it so that .RESET directives caught in the parser can properly reset BA and PO. This should hopefully make it a little less likely that BA and PO end up desynced cuz we missed a relevant directive, though the possibility is still there (the parser can't currently handle .BA/.PO being set to label locations, nor can it handle .GOTO statements period, which is going to throw things off).

I don't know that I've stated this before necessarily, but it should be made clear, this project is first and foremost meant to act on hex-code Gecko and PPC, not so much on GCTRM formatted input. I'll do what I can to better support things like .BA/PO<-LABEL directives and condition block tracking in the future (all provided I have the time to do so), but for now these limitations are gonna be what they're gonna be.